REFACTOR #428 : ApprovedStatus enum 의 폴더 위치 변경
