+++ src/main/java/com/lokoko/domain/user/application/usecase/AdminUsecase.java
@@ -4,7 +4,7 @@ import com.lokoko.domain.campaign.api.dto.response.CampaignBasicResponse;
 import com.lokoko.domain.campaign.application.service.CampaignGetService;
 import com.lokoko.domain.campaign.domain.repository.CampaignRepository;
 import com.lokoko.domain.user.api.dto.request.ApproveCampaignIdsRequest;
-import com.lokoko.domain.user.api.dto.request.ApprovedStatus;
+import com.lokoko.domain.user.domain.entity.enums.ApprovedStatus;
 import com.lokoko.domain.user.api.dto.request.CampaignModifyRequest;
 import com.lokoko.domain.user.api.dto.request.DeleteCampaignIdsRequest;
 import com.lokoko.domain.user.api.dto.response.AdminCampaignListResponse;
